knights/requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.1
django-better-admin-arrayfield==1.4.0
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-crispy-forms==1.9.2
django-forms-bootstrap==3.1.0
django-js-asset==1.2.2
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.3.1
uWSGI==2.0.19.1
